proxy/requirements.txt
starlette>=0.27.0
uvicorn>=0.23.0
httpx>=0.24.0
python-dotenv>=1.0.1
